move 'eye' to twodim_base
